refactor: paintTools 패키지 이동
